Início da refatoração da passagem de dados entre activities. Criação da BaseActivity

+++ domain/src/main/java/bitbox/project/domain/model/user/User.kt
@@ -4,7 +4,7 @@ data class User(
     val erro: Int? = 0,
     val idUsuario: Int,
     val mensagem: String = "",
-    val saldo: Int,
+    val saldo: Double,
     val saldoPendente: Int,
     val transacoesPendentes: List<Any>
 )
